Clarification & Compensation
Fixed a poor implementation that was conflating step distance with the
number of steps per movement.

Changed semantics of several variables and behaviors for clarity.

While not perfect, now attempts to compensate for step distances that
are not divisible into the cell size. Still recommend having cell size
divisible by step distance.